refactor: unify membership and stable_set
